+++ SBBDesignSystemMobileSwiftUI/Views/SBBOnboardingWrapper/SBBCardView.swift
@@ -4,15 +4,14 @@
 
 import SwiftUI
 
-public struct SBBCardView<Card: Equatable>: View {
+public struct SBBCardView: View {
     
     private let image: Image
     private let title: Text
     private let text: Text
     private let size: CGSize?
     private let canBeTried: Bool
-    private let nextCard: Card?
-    private let onNext: (Card?) -> Void
+    private let onNext: () -> Void
     @Binding private var showTrySheet: Bool
     
     @Environment(\.horizontalSizeClass) private var horizontalSizeClass
@@ -24,14 +23,13 @@ public struct SBBCardView<Card: Equatable>: View {
     
     private let scrollViewHeight: CGFloat?
     
-    public init(image: Image, title: Text, text: Text, size: CGSize? = nil, showTrySheet: Binding<Bool>, nextCard: Card?, onNext: @escaping (Card?) -> Void) {
+    public init(image: Image, title: Text, text: Text, size: CGSize? = nil, showTrySheet: Binding<Bool>, onNext: @escaping () -> Void) {
         self.image = image
         self.title = title
         self.text = text
         self.size = size
         self.canBeTried = true
         self._showTrySheet = showTrySheet
-        self.nextCard = nextCard
         self.onNext = onNext
         
         self.contentWidthLandscape = size?.width != nil ? 3 * size!.width / 4 : nil
@@ -39,14 +37,13 @@ public struct SBBCardView<Card: Equatable>: View {
         self.scrollViewHeight = size?.width != nil ? size!.height - 54 - 16 : nil
     }
     
-    public init(image: Image, title: Text, text: Text, size: CGSize? = nil, nextCard: Card?, onNext: @escaping (Card?) -> Void) {
+    public init(image: Image, title: Text, text: Text, size: CGSize? = nil, onNext: @escaping () -> Void) {
         self.image = image
         self.title = title
         self.text = text
         self.size = size
         self.canBeTried = false
         self._showTrySheet = .constant(false)
-        self.nextCard = nextCard
         self.onNext = onNext
         
         self.contentWidthLandscape = size?.width != nil ? 3 * size!.width / 4 : nil
@@ -74,7 +71,7 @@ public struct SBBCardView<Card: Equatable>: View {
     private var nextButton: some View {
         Button(action: {
             withAnimation {
-                onNext(nextCard)
+                onNext()
             }
             UIAccessibility.post(notification: .screenChanged, argument: nil)   // reset voiceover focus (to the current card)
         }) {
@@ -234,9 +231,9 @@ public struct SBBCardView<Card: Equatable>: View {
 }
 
 #Preview("Without try function") {
-    SBBCardView<Int>(image: Image("Onboarding_Card1"), title: Text("Title"), text: Text("This is the content of the card"), nextCard: nil, onNext: { _ in })
+    SBBCardView(image: Image("Onboarding_Card1"), title: Text("Title"), text: Text("This is the content of the card"), onNext: { })
 }
 
 #Preview("With try function") {
-    SBBCardView<Int>(image: Image("Onboarding_Card1"), title: Text("Title"), text: Text("This is the content of the card"), showTrySheet: .constant(false), nextCard: nil, onNext: { _ in })
+    SBBCardView(image: Image("Onboarding_Card1"), title: Text("Title"), text: Text("This is the content of the card"), showTrySheet: .constant(false), onNext: { })
 }