+++ SBBDesignSystemMobileSwiftUIDemo/Views/DetailViews/OnboardingViewDemo.swift
@@ -85,7 +85,7 @@ struct OnboardingViewDemo: View {
                         Alert(title: Text("Custom Action"), message: Text("This alert is presented as a custom executable action on card disappear."), dismissButton: .default(Text("Got it!")))
                     }
             } else {
-                SBBOnboardingWrapper(state: $overviewState, currentCard: $currentOnboardingCardIndex, currentCardIndex: $currentOnboardingCardIndex, nbCards: viewModel.numberOfCards, startView: {
+                SBBOnboardingWrapper(state: overviewState, currentCard: currentOnboardingCardIndex, currentCardIndex: currentOnboardingCardIndex, nbCards: viewModel.numberOfCards, startView: {
                     SBBTitleView(image: Image("Onboarding_Luc"), title: Text("Willkommen"), subtitle: Text("Willkommen zum Rundgang."), buttonView: {
                         Button(action: {
                             withAnimation {
@@ -107,15 +107,11 @@ struct OnboardingViewDemo: View {
                         }
                         .buttonStyle(SBBPrimaryButtonStyle())
                     })
-                }, cardBuilder: { card in
+                }, cardBuilder: { card, geometrySize in
                     if card % 2 == 1 {
-                        GeometryReader { geometry in
-                            createCardView(card, viewModel.withCustomCard, viewModel.withCustomAction, viewModel.withCustomButton, geometry.size)
-                        }
+                        createCardView(card, viewModel.withCustomCard, viewModel.withCustomAction, viewModel.withCustomButton, geometrySize)
                     } else {
-                        GeometryReader { geometry in
-                        createCardView(card, viewModel.withCustomCard, viewModel.withCustomAction, viewModel.withCustomButton, geometry.size)
-                        }
+                        createCardView(card, viewModel.withCustomCard, viewModel.withCustomAction, viewModel.withCustomButton, geometrySize)
                     }
                 })
                 .sheet(isPresented: $showTrySheet, content: {
@@ -183,7 +179,7 @@ struct OnboardingViewDemo: View {
                 
                 Button(action: {
                     currentOnboardingCardIndex += 1
-                    if card == viewModel.numberOfCards {
+                    if card == viewModel.numberOfCards - 1 {
                         overviewState = .end
                     }
                 }) {
@@ -204,16 +200,15 @@ struct OnboardingViewDemo: View {
                 Spacer()
             }
             .sbbScreenPadding()
-            .transition(.backslide)
             .onDisappear {
                 if customAction {
                     showingAlert = true
                 }
             }
         } else if customButton {
-            SBBCardView(image: Image("Onboarding_Card2"), title: Text("Card"), text: Text("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."), size: geometrySize, showTrySheet: $showTrySheet, nextCard: card+1, onNext: { card in
+            SBBCardView(image: Image("Onboarding_Card2"), title: Text("Card"), text: Text("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."), size: geometrySize, showTrySheet: $showTrySheet, onNext: {
                 currentOnboardingCardIndex += 1
-                if card == viewModel.numberOfCards {
+                if card == viewModel.numberOfCards - 1 {
                     overviewState = .end
                 }
             })
@@ -223,9 +218,9 @@ struct OnboardingViewDemo: View {
                 }
             }
         } else {
-            SBBCardView(image: Image("Onboarding_Card2"), title: Text("Card"), text: Text("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."), size: geometrySize, nextCard: card+1, onNext: { card in
+            SBBCardView(image: Image("Onboarding_Card2"), title: Text("Card"), text: Text("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."), size: geometrySize, onNext: {
                 currentOnboardingCardIndex += 1
-                if card == viewModel.numberOfCards {
+                if card == viewModel.numberOfCards - 1 {
                     overviewState = .end
                 }
             })