+++ ReleaseNotes.md
@@ -1,5 +1,11 @@
 # Release Notes SBBDesignSystemMobileSwiftUI for iOS & SwiftUI
 
+## v1.2.4 - 3 July 2025
+### Improvement 
+* SBBPagination: `currentPageIndex` is not a `Binding`, as it is never changed.
+* SBBCardView: `nextCard` attribute is removed, as it is not used. `onNext` does not take a `nextCard` attribute.
+* SBBOnboardingWrapper: `GeometryReader` added for the card builder, so that we get the size from it. No `Binding` attributes anymore: no state management is done in the wrapper, it simply used to display the views.
+
 ## v1.2.3 - 1 July 2025
 ### Features
 * New component: